feat: add routing and pages
